Refactor point and direction to their own modules
