+++ src/bin/golf-peaks.rs
@@ -1,109 +1,7 @@
 use anyhow::{anyhow, Result};
-use std::{
-    io::{BufRead, Read},
-    ops::{Add, Mul, Sub},
-};
+use std::io::{BufRead, Read};
 
-use solver::{Solver, State};
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, PartialOrd, Ord)]
-struct Point {
-    x: isize,
-    y: isize,
-}
-
-impl Point {
-    #[allow(dead_code)]
-    fn manhattan_distance(&self, other: Point) -> isize {
-        (self.x - other.x).abs() + (self.y - other.y).abs()
-    }
-}
-
-impl Add<Point> for Point {
-    type Output = Point;
-
-    fn add(self, other: Point) -> Point {
-        Point {
-            x: self.x + other.x,
-            y: self.y + other.y,
-        }
-    }
-}
-
-impl Sub<Point> for Point {
-    type Output = Point;
-
-    fn sub(self, other: Point) -> Point {
-        Point {
-            x: self.x - other.x,
-            y: self.y - other.y,
-        }
-    }
-}
-
-impl Mul<isize> for Point {
-    type Output = Point;
-
-    fn mul(self, other: isize) -> Point {
-        Point {
-            x: self.x * other,
-            y: self.y * other,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-enum Direction {
-    Up,
-    Down,
-    Left,
-    Right,
-}
-
-impl From<Direction> for Point {
-    fn from(direction: Direction) -> Point {
-        match direction {
-            Direction::Up => Point { x: 0, y: -1 },
-            Direction::Down => Point { x: 0, y: 1 },
-            Direction::Left => Point { x: -1, y: 0 },
-            Direction::Right => Point { x: 1, y: 0 },
-        }
-    }
-}
-
-impl Direction {
-    fn all() -> Vec<Direction> {
-        vec![
-            Direction::Up,
-            Direction::Down,
-            Direction::Left,
-            Direction::Right,
-        ]
-    }
-
-    fn flip(&self) -> Direction {
-        match self {
-            Direction::Up => Direction::Down,
-            Direction::Down => Direction::Up,
-            Direction::Left => Direction::Right,
-            Direction::Right => Direction::Left,
-        }
-    }
-}
-
-impl TryFrom<&str> for Direction {
-    type Error = anyhow::Error;
-
-    fn try_from(value: &str) -> Result<Self> {
-        match value {
-            "up" => Ok(Direction::Up),
-            "down" => Ok(Direction::Down),
-            "left" => Ok(Direction::Left),
-            "right" => Ok(Direction::Right),
-            _ => Err(anyhow!("Invalid direction: {value}")),
-        }
-    }
-}
+use solver::{Solver, State, direction::Direction, point::Point};
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 enum AngleType {
