+++ src/point.rs
@@ -1,0 +1,47 @@
+use std::ops::{Add, Mul, Sub};
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, PartialOrd, Ord)]
+pub struct Point {
+    pub x: isize,
+    pub y: isize,
+}
+
+#[allow(dead_code)]
+impl Point {
+    pub fn manhattan_distance(&self, other: Point) -> isize {
+        (self.x - other.x).abs() + (self.y - other.y).abs()
+    }
+}
+
+impl Add<Point> for Point {
+    type Output = Point;
+
+    fn add(self, other: Point) -> Point {
+        Point {
+            x: self.x + other.x,
+            y: self.y + other.y,
+        }
+    }
+}
+
+impl Sub<Point> for Point {
+    type Output = Point;
+
+    fn sub(self, other: Point) -> Point {
+        Point {
+            x: self.x - other.x,
+            y: self.y - other.y,
+        }
+    }
+}
+
+impl Mul<isize> for Point {
+    type Output = Point;
+
+    fn mul(self, other: isize) -> Point {
+        Point {
+            x: self.x * other,
+            y: self.y * other,
+        }
+    }
+}
